+++ Modules/booking/Controller/BookingquantitiesController.php
@@ -25,6 +25,7 @@ class BookingquantitiesController extends BookingsupsabstractController
         $this->invoicable = true;
         $this->mandatoryFields = true;
         $this->hasDuration = false;
+        $this->hasChoices = true;
         $this->formUrl = "bookingquantities";
     }