+++ db/upgrade/1661781152_add_ac_anticorps_reactivity_column.php
@@ -2,8 +2,10 @@
 require_once 'Framework/Model.php';
 require_once 'Framework/Configuration.php';
 # Upgrade: add reactivity column to ac_anticorps
-class CoreUpgradeDB1661781152 extends Model {
-  public function run(){
+class CoreUpgradeDB1661781152 extends Model
+{
+  public function run()
+  {
     Configuration::getLogger()->info("[db][upgrade] Apply add reactivity column to table ac_anticorps");
     $sql = 'ALTER TABLE ac_anticorps ADD COLUMN reactivity varchar(30) NOT NULL DEFAULT ""';
     $this->runRequest($sql);
@@ -12,4 +14,3 @@ class CoreUpgradeDB1661781152 extends Model {
 }
 $db = new CoreUpgradeDB1661781152();
 $db->run();
-?>